removed -dump from Makefile
Removed broken test cases from RegressionTesting.  PA5funcexampleCalleeSaved.java.s
was an example that I was in the middle of doing some AVR rewriting by hand.  It never
came to fruition.  PA5NegCompares is causing the code generator and/or the simulator
problems.  I moved it to the MeggyRef/RegressionTesting.